assets/Asteroid le debut de la fin/Asteroid/requirements.txt
cocos2d==0.6.5
future==0.16.0
pygame==1.9.4
pyglet==1.3.0
six==1.11.0
